+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.11
 - Ability to download deployed process
 - Added locking objects capabilities
+- Added removing color mapping form a report definition
 
 ## 0.6.10
 - Fixed client default missing in ProjectMetadata